+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/Verifier.kt
@@ -17,11 +17,6 @@ import kotlinx.serialization.json.JsonObject
 interface Verifier {
 
     /**
-     * The public key for this agent, i.e. the one used to validate the audience of a VP against
-     */
-    val keyMaterial: KeyMaterial
-
-    /**
      * Set the revocation list to use for validating VCs (from [Issuer.issueRevocationListCredential])
      */
     fun setRevocationList(it: String): Boolean